requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib